[DDH-13] Create container; Change file type to jsx
