+++ src/app/page.js
@@ -1,4 +1,4 @@
-import MapFilter from "@/components/MapFilter/MapFilter";
+import { MapFilter } from "@/containers";
 
 export default function Home() {
   return <div className="w-full h-full">
